refactor: 좌석, 날짜를 포함하여 하나의 엔티티로 생성
